transform: Add SegmentedColumnTransformer
The main use case for this transformer is to enable the application of specified
groups of feature functions to specified columns of data, e.g. when dealing with
heterogeneous data.

The SegmentedColumnTransformer is derived from the sklearn ColumnTransformer
and adapted to be used inside a Pype object after a segment transformation.

The adaption mainly consists of:
- adapt the notation of a column (ColumnTransformer iterates over the second
  dimension, segmented data must be iterated over the third dimension).
- disable "drop" and "passthrough" transform options for simplicity and drop
  non-specified columns by default

Note: SegmentedColumnTransformer does not support contextual data.