transform: Add patch_sampler
This functions dynamically patches an imbalanced-learn Sampler
transformer to be usable inside a seglearn Pype. It ensures that the
objects created from this metaclass are pickable.

Additionally, shuffling is implemented for imbalanced-learn samplers.
The reason is that imbalanced-learn sorts the output by classes. This is
problematic when splitting the resampled data (e.g. using the validation
split from the Keras fit function).

Finally, calling repr() on a dynamically patched sampler will return the
parameters of the imbalanced-learn base class along with the additional
parameters introduced in the PickableSampler class (shuffle and
random_state).